+++ snmplib/snmp_enum.c
@@ -161,7 +161,7 @@ se_store_enum_list(struct snmp_enum_list *new_list,
          *   then save the line so far, and start again.
          */
 	len = sizeof(line) - strlen(line);
-	if (strlen(buf) > len) {
+	if ((int)strlen(buf) > len) {
 	    read_config_store(type, line);
             snprintf(line, sizeof(line), "enum %s", token);
 	    len = sizeof(line);
